[Android Target] move type adapters to android-client
