+++ kotlin-generator/src/android-client.ts
@@ -11,15 +11,11 @@ import android.util.Base64
 import com.google.gson.*
 import com.google.gson.reflect.TypeToken
 import com.google.gson.annotations.JsonAdapter
-import com.google.gson.stream.JsonReader
-import com.google.gson.stream.JsonWriter
 import io.sdkgen.runtime.SdkgenHttpClient
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Deferred
 import kotlinx.coroutines.Dispatchers.IO
 import kotlinx.coroutines.async
-import java.lang.reflect.Type
-import java.text.SimpleDateFormat
 import java.util.*
 
 inline fun <reified T> Gson.fromJson(json: String) =
@@ -36,89 +32,6 @@ class ApiClient(
 ) : SdkgenHttpClient(baseUrl, applicationContext, defaultTimeoutMillis) {`;
 
     code += `
-    private class ByteArrayDeserializer : JsonDeserializer<ByteArray> {
-        @Throws(JsonParseException::class)
-        override fun deserialize(
-            element: JsonElement,
-            arg1: Type?,
-            arg2: JsonDeserializationContext?
-        ): ByteArray {
-            try {
-                return Base64.decode(element.asString, Base64.DEFAULT)
-            } catch (e: Exception) {
-                throw JsonIOException(e.message)
-            }
-        }
-    }
-
-    private class DateTimeAdapter: TypeAdapter<Calendar>() {
-        companion object {
-            val sdf = SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSS", Locale.US).apply { 
-                this.timeZone = TimeZone.getTimeZone("UTC")
-            }
-        }
-        
-        override fun write(out: JsonWriter?, value: Calendar?) {
-            out?.let {
-                if (value == null) {
-                    it.nullValue()
-                    return
-                }
-
-                val dateTimeString = sdf.format(value.time)
-                it.value(dateTimeString)
-            }
-        }
-
-        override fun read(reader: JsonReader?): Calendar? {
-            reader?.let {
-                val dateTimeString = it.nextString()
-
-                try {
-                    return Calendar.getInstance().apply { time = sdf.parse(dateTimeString)!! }
-                } catch (e: Exception) {
-                    throw JsonIOException(e.message)
-                }
-            } ?: run { 
-                return null
-            }
-        }
-    }
-
-    private class DateAdapter: TypeAdapter<Calendar>() {
-        companion object {
-            val sdf = SimpleDateFormat("yyyy-MM-dd", Locale.US).apply {
-                this.timeZone = TimeZone.getTimeZone("UTC")
-            }
-        }
-
-        override fun write(out: JsonWriter?, value: Calendar?) {
-            out?.let {
-                if (value == null) {
-                    it.nullValue()
-                    return
-                }
-
-                val dateTimeString = sdf.format(value.time)
-                it.value(dateTimeString)
-            }
-        }
-
-        override fun read(reader: JsonReader?): Calendar? {
-            reader?.let {
-                val dateTimeString = it.nextString()
-                
-                try {
-                    return Calendar.getInstance().apply { time = sdf.parse(dateTimeString)!! }
-                } catch (e: Exception) {
-                    throw JsonIOException(e.message)
-                }
-            } ?: run {
-                return null
-            }
-        }
-    }
-
     private val gson = GsonBuilder()
         .registerTypeAdapter(object : TypeToken<ByteArray>() {}.type, ByteArrayDeserializer())
         .create()\n\n`;
